requirements.txt
# packages in environment at /Users/greenelaia/opt/anaconda3/envs/ch03_env:
#
# Name                    Version                   Build  Channel
altair                    5.2.0                    pypi_0    pypi
annotated-types           0.6.0                    pypi_0    pypi
anyio                     4.3.0                    pypi_0    pypi
attrs                     23.2.0                   pypi_0    pypi
blinker                   1.7.0                    pypi_0    pypi
bzip2                     1.0.8                h1de35cc_0  
ca-certificates           2023.12.12           hecd8cb5_0  
cachetools                5.3.2                    pypi_0    pypi
certifi                   2024.2.2                 pypi_0    pypi
charset-normalizer        3.3.2                    pypi_0    pypi
click                     8.1.7                    pypi_0    pypi
distro                    1.9.0                    pypi_0    pypi
exceptiongroup            1.2.0                    pypi_0    pypi
gitdb                     4.0.11                   pypi_0    pypi
gitpython                 3.1.42                   pypi_0    pypi
gtts                      2.5.1                    pypi_0    pypi
h11                       0.14.0                   pypi_0    pypi
httpcore                  1.0.3                    pypi_0    pypi
httpx                     0.26.0                   pypi_0    pypi
idna                      3.6                      pypi_0    pypi
importlib-metadata        7.0.1                    pypi_0    pypi
jinja2                    3.1.3                    pypi_0    pypi
jsonschema                4.21.1                   pypi_0    pypi
jsonschema-specifications 2023.12.1                pypi_0    pypi
libffi                    3.4.4                hecd8cb5_0  
markdown-it-py            3.0.0                    pypi_0    pypi
markupsafe                2.1.5                    pypi_0    pypi
mdurl                     0.1.2                    pypi_0    pypi
ncurses                   6.4                  hcec6c5f_0  
numpy                     1.26.4                   pypi_0    pypi
openai                    1.12.0                   pypi_0    pypi
openssl                   3.0.13               hca72f7f_0  
packaging                 23.2                     pypi_0    pypi
pandas                    2.2.0                    pypi_0    pypi
pillow                    10.2.0                   pypi_0    pypi
pip                       23.3.1                   pypi_0    pypi
protobuf                  4.25.3                   pypi_0    pypi
pyarrow                   15.0.0                   pypi_0    pypi
pydantic                  2.6.1                    pypi_0    pypi
pydantic-core             2.16.2                   pypi_0    pypi
pydeck                    0.8.1b0                  pypi_0    pypi
pydub                     0.25.1                   pypi_0    pypi
pygments                  2.17.2                   pypi_0    pypi
python                    3.10.13              h5ee71fb_0  
python-dateutil           2.8.2                    pypi_0    pypi
pytz                      2024.1                   pypi_0    pypi
readline                  8.2                  hca72f7f_0  
referencing               0.33.0                   pypi_0    pypi
requests                  2.31.0                   pypi_0    pypi
rich                      13.7.0                   pypi_0    pypi
rpds-py                   0.18.0                   pypi_0    pypi
setuptools                68.2.2                   pypi_0    pypi
six                       1.16.0                   pypi_0    pypi
smmap                     5.0.1                    pypi_0    pypi
sniffio                   1.3.0                    pypi_0    pypi
sqlite                    3.41.2               h6c40b1e_0  
streamlit                 1.31.1                   pypi_0    pypi
streamlit-audiorecorder   0.0.4                    pypi_0    pypi
tenacity                  8.2.3                    pypi_0    pypi
tk                        8.6.12               h5d9f67b_0  
toml                      0.10.2                   pypi_0    pypi
toolz                     0.12.1                   pypi_0    pypi
tornado                   6.4                      pypi_0    pypi
tqdm                      4.66.2                   pypi_0    pypi
typing-extensions         4.9.0                    pypi_0    pypi
tzdata                    2024.1                   pypi_0    pypi
tzlocal                   5.2                      pypi_0    pypi
urllib3                   2.2.1                    pypi_0    pypi
validators                0.22.0                   pypi_0    pypi
wheel                     0.41.2                   pypi_0    pypi
xz                        5.4.5                h6c40b1e_0  
zipp                      3.17.0                   pypi_0    pypi
zlib                      1.2.13               h4dc903c_0  
